some smol pixy changes. Today Marcel and I made wires for the pixy to rio connection.
About the wires, so right now there's a 6 pin male end that you plug into the Pixy's ribbon cable end. The other end you use MXP pins to connect from, refer pinouts as necessary.

In software you NEED to go to PixyMon and set it to Arduino with ICSP mode, NOT SPI with SS mode. Right now we guess it should work like this and will default to be CS0 to the Rio, and a gyro can be CS1 using its jumper pins.

Also i forgot to instantiate timer because i programn't